% --- compareGroups-internal.Rd ---

\name{compareGroups-internal}
\alias{compareGroups-internal}
\alias{chisq.test2}
\alias{combn2}
\alias{compare.i}
\alias{descrip}
\alias{flip}
\alias{format2}
\alias{norm.plot}
\alias{signifdec.i}
\alias{signifdec}
\alias{summ.i}
\alias{table.i}


\title{Internal compareGroups functions}
\description{Internal compareGroups functions}
\usage{
chisq.test2(obj)
combn2(x)
compare.i(x, y, selec.i, method.i, alpha, min.dis, max.xlev, varname)
descrip(x, y, method = "param")
flip(x)
format2(x, digits = NULL, ...)
formula.compareGroups(x, ...) 
norm.plot(var,var.label,z,n.breaks,plot.new=TRUE)
signifdec.i(x, digits)
signifdec(x, digits)
summ.i(x)
table.i(x, hide.i, digits, type, varname, show.all, show.p.trend, show.p.mul, show.n)
update.formula2(old, new, ...) 
}

\details{These are not to be called by the user}
\keyword{internal}
