\name{compareGroups-package}
\alias{compareGroups-package}
\alias{compareGroups-package}
\docType{package}
\title{
Descriptives analysis by groups
}
\description{
A package which quickly and efficiently generates descriptives by groups (bivariate tables) in plain text, or as LATEX instructions. Depending on the nature of the variable, different descriptive statistics can be calculated (mean, median, frequencies) and different tests are computed as appropriate (t-test, ANOVA, Kruskall-Wallis, Fisher,...). In order to facilitate construction of the bivariate table, we have implemented a graphical user interface.
}
\details{
\tabular{ll}{
Package: \tab compareGroups\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2010-06-26\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
Main functions: \code{\link{compareGroups}}, \code{\link{createTable}}, \code{\link{export2latex}}, \code{\link{export2latex}}, \code{\link{export2csv}}, \code{\link{cGroupsGUI}}    
}

\author{
Héctor Sanz <hsanz@imim.es>, Isaac Subirana <isubirana@imim.es> and Joan Vila <jvila@imim.es>

Maintainer: Isaac Subirana <isubirana@imim.es>
}

\keyword{ package }

